Use better local testing approach
The new testing approach makes it easier to test the theme locally,
without the need to build the theme gem through the Rakefile every time
a change is made to the theme's files, which happens too often.

So, a separate test directory was created with the following files:
  - _config.yml: a modified version of the original _config.yml
  - Gemfile: uses the gemspec method from root directory.
  - .gitignore

The old approach's files were removed:
  - Rakefile
  - docs/_config.dev.yml

The following files were updated; to work with the new approach:
  - _includes/navbar.html
  - docs/pages/docs.md
  - docs/search.json

This approach has only one bug which is that the blog permalink is set
to docs/blog/ instead of blog/; to avoid breaking the pagination.